% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{birp_from_command_line}
\alias{birp_from_command_line}
\title{Creating a Birp Object based on output files of command-line tool}
\usage{
birp_from_command_line(path)
}
\arguments{
\item{path}{The path where all the output files of birp are located}
}
\value{
An object of class birp
}
\description{
This function creates a birp object by reading the output files of the command-line tool
}
\examples{
est <- birp_from_command_line(file.path(system.file("extdata", package = "birp")))
}
