% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{.addTextSingleGamma.birp}
\alias{.addTextSingleGamma.birp}
\title{Function to add text box to plot denoting P(gamma > 0 | n) or P(gamma < 0 | n) for single gammas}
\usage{
.addTextSingleGamma.birp(x)
}
\arguments{
\item{x}{A birp object}
}
\value{
No return value, called for side effects.
}
\description{
Function to add text box to plot denoting P(gamma > 0 | n) or P(gamma < 0 | n) for single gammas
}
\keyword{internal}
