% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{.addTextSingleGammaDelta.birp}
\alias{.addTextSingleGammaDelta.birp}
\title{Function to add text box to plot denoting P(gamma > 0 | n) or P(gamma < 0 | n) for single gammas (same for Delta)}
\usage{
.addTextSingleGammaDelta.birp(post, param_name = "gamma")
}
\arguments{
\item{post}{The posterior probabilities (gamma or Delta)}

\item{param_name}{The parameter name to show}
}
\value{
No return value, called for side effects.
}
\description{
Function to add text box to plot denoting P(gamma > 0 | n) or P(gamma < 0 | n) for single gammas (same for Delta)
}
\keyword{internal}
