% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{.addToList.birp}
\alias{.addToList.birp}
\title{Function to convert an argument to a string and add it to a list if necessary}
\usage{
.addToList.birp(options, name, x)
}
\arguments{
\item{options}{A list where x should be added to}

\item{name}{A string specifying the name of the argument}

\item{x}{An R object to be added to the list}
}
\value{
An updated list options
}
\description{
Function to convert an argument to a string and add it to a list if necessary
}
\keyword{internal}
