% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{.calculateRho.birp}
\alias{.calculateRho.birp}
\title{Function to calculate rho}
\usage{
.calculateRho.birp(times, times_of_change)
}
\arguments{
\item{times}{An integer or numeric vector with time points}

\item{times_of_change}{An integer or numeric vector with times of change}
}
\value{
A numeric matrix containing rho for each time point and epoch
}
\description{
Function to calculate rho
}
\keyword{internal}
