% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{.createObjBirp.birp}
\alias{.createObjBirp.birp}
\title{Function to create an object of type birp}
\usage{
.createObjBirp.birp(
  data,
  meanVar,
  trace,
  gamma,
  Delta,
  timepoints,
  timesOfChange,
  rate_design,
  step_design,
  CI_groups,
  state
)
}
\arguments{
\item{data}{An object of type \link{birp_data}, corresponding to filtered data used for inference}

\item{meanVar}{A data frame containing the posterior mean and variance of all parameters}

\item{trace}{A data frame containing the MCMC trace of all parameters}

\item{gamma}{A data frame containing the posterior probabilities regarding gamma}

\item{Delta}{A data frame containing the posterior probabilities regarding Delta}

\item{timepoints}{An integer vector containing the timepoints at which counts were obtained}

\item{timesOfChange}{A numeric or integer vector specifying the times of change}

\item{rate_design}{A matrix specifying the BACI configuration for the rates of change (gamma, see details).}

\item{step_design}{A matrix specifying the BACI configuration for the step changes (Delta, see details).}

\item{CI_groups}{A character vector specifying the names of the control-intervention (CI) group}

\item{state}{A data frame containing the posterior mean values of all parameters inferred by birp}
}
\value{
An object of type birp
}
\description{
Function to create an object of type birp
}
\details{
The `rate_design` and `step_design` matrices define a Before-After Control-Impact experimental design for the rates of change (gamma) and the step changes (Delta), respectively, with the following format:
- Each **row** represents a group (e.g., Control or Intervention). The **first column** specifies the group name (e.g. 'Control' or 'Intervention').
- Each **column after the first** represents a different epoch. The numbers in these columns indicate which change parameter (\eqn{\gamma} or \eqn{\Delta}) to assign for each group and epoch.
For example, BACI = matrix(c("A", "B", 1, 1, 1, 2), nrow = 2) corresponds to a canonical BACI design where the first row represents the control group (A) and the second row represents the intervention group (B). 
Please see the vignette for more examples.
}
\keyword{internal}
