% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp_data.R
\name{.getDataAllMethods.birp_data}
\alias{.getDataAllMethods.birp_data}
\title{This function reads data for all methods into a list}
\usage{
.getDataAllMethods.birp_data(out, what, res)
}
\arguments{
\item{out}{A string corresponding to the output prefix}

\item{res}{A list of dataframes from Rcpp}
}
\value{
An instance of birp_data
}
\description{
This function reads data for all methods into a list
}
\keyword{internal}
