% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{.getEpochStarts.birp}
\alias{.getEpochStarts.birp}
\title{Function to get the start of each epoch, including for a hypothetical epoch after the last epoch}
\usage{
.getEpochStarts.birp(times, times_of_change)
}
\arguments{
\item{times}{An integer or numeric vector with time points}

\item{times_of_change}{An integer or numeric vector with times of change}
}
\value{
An integer or numeric vector with starting points for each epoch
}
\description{
Function to get the start of each epoch, including for a hypothetical epoch after the last epoch
}
\keyword{internal}
