% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{.getLabelGamma.birp}
\alias{.getLabelGamma.birp}
\title{Function to generate a nice axis label with greek gamma and subscript}
\usage{
.getLabelGamma.birp(x, index)
}
\arguments{
\item{index}{The index in the subscript}
}
\value{
A string
}
\description{
Function to generate a nice axis label with greek gamma and subscript
}
\keyword{internal}
