% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{.openFile.birp}
\alias{.openFile.birp}
\title{Function to open a file and generate error message if it was not found}
\usage{
.openFile.birp(
  path,
  files,
  patterns,
  sep = "\\t",
  header = TRUE,
  mustExist = TRUE
)
}
\arguments{
\item{path}{A file path.}

\item{files}{A vector of character strings corresponding to file names found in the path.}

\item{patterns}{A vector of patterns to search for within 'files'.}

\item{sep}{The field separator character of the input file}

\item{mustExist}{Logical. If \code{TRUE}, an exception is raised if the file does not exist}
}
\value{
A file connection.
}
\description{
Function to open a file and generate error message if it was not found
}
\keyword{internal}
