% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{.openPosteriorPlot.birp}
\alias{.openPosteriorPlot.birp}
\title{Function to open an empty plot for plotting the posterior probabilities}
\usage{
.openPosteriorPlot.birp(
  xlim,
  ylim,
  xlab,
  ylab,
  shadingIncrease,
  shadingDecrease,
  lineAtZero,
  ...
)
}
\arguments{
\item{xlim}{The x-limits (x1, x2) of the plot}

\item{ylim}{The y-limits (y1, y2) of the plot}

\item{xlab}{Name of x axis}

\item{ylab}{Name of y axis}

\item{shadingIncrease}{Shading color for the range gamma > 0. If \code{NA}, shading is omitted}

\item{shadingDecrease}{Shading color for the range gamma < 0. If \code{NA}, shading is omitted}

\item{lineAtZero}{If \code{TRUE}, adds a dashed line indicating 0.}
}
\value{
No return value, called for side effects.
}
\description{
Function to open an empty plot for plotting the posterior probabilities
}
\keyword{internal}
