% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{.parsePosteriorGammaDelta.birp}
\alias{.parsePosteriorGammaDelta.birp}
\title{Function to parse posterior results of gamma and Delta}
\usage{
.parsePosteriorGammaDelta.birp(param_name, meanVar, trace, posterior_summary)
}
\arguments{
\item{param_name}{A string defining the parameter name (gamma or Delta)}

\item{meanVar}{A data frame containing the posterior mean and variance of all parameters}

\item{trace}{A data frame containing the MCMC trace of all parameters}

\item{posterior_summary}{A data frame containing the posterior probabilities}
}
\value{
A list with relevant posterior statistics
}
\description{
Function to parse posterior results of gamma and Delta
}
\keyword{internal}
