% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{.plotShadingPolygon.birp}
\alias{.plotShadingPolygon.birp}
\title{Function to add a hatched polygon to a plot}
\usage{
.plotShadingPolygon.birp(shading, left, right)
}
\arguments{
\item{shading}{Shading color. If \code{NA}, shading is omitted}

\item{left}{An integer indicating the left-most value on the x-axis}

\item{right}{An integer indicating the right-most value on the x-axis}
}
\value{
No return value, called for side effects.
}
\description{
Function to add a hatched polygon to a plot
}
\keyword{internal}
