% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp_data.R
\name{.validate.birp_data}
\alias{.validate.birp_data}
\title{This function check if a birp_data object is valid}
\usage{
.validate.birp_data(data)
}
\arguments{
\item{data}{A list of dataframes to validate}
}
\value{
No return value, called for side effects.
}
\description{
This function check if a birp_data object is valid
}
\keyword{internal}
