% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{print.birp}
\alias{print.birp}
\title{Print a birp object}
\usage{
\method{print}{birp}(x, ...)
}
\arguments{
\item{x}{A \code{birp} object.}

\item{...}{Additional arguments passed to internal methods (currently unused).}
}
\value{
Invisibly returns the input \code{x}, called for side effects.
}
\description{
Prints a summary of the estimated parameters from a \code{birp} model.
}
\examples{
data <- simulate_birp()
est <- birp(data)
print(est)

}
\seealso{
\code{\link{birp}}
}
