% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp_data.R
\name{print.birp_data}
\alias{print.birp_data}
\title{Print a birp_data object
Prints a summary of a \code{birp_data} object, including the number of methods, locations, control-intervention (CI) groups, and timepoints, as well as the names or identifiers for each.}
\usage{
\method{print}{birp_data}(x, ...)
}
\arguments{
\item{x}{A \code{birp_data} object to be printed.}

\item{...}{Additional arguments passed to function.}
}
\value{
No return value; this function is called for its side effects (printing to console).
}
\description{
Print a birp_data object
Prints a summary of a \code{birp_data} object, including the number of methods, locations, control-intervention (CI) groups, and timepoints, as well as the names or identifiers for each.
}
\examples{
data <- simulate_birp()
print(data)
}
