% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{summary.birp}
\alias{summary.birp}
\title{Summary method for birp objects}
\usage{
\method{summary}{birp}(object, ...)
}
\arguments{
\item{object}{A \code{birp} object.}

\item{...}{Additional arguments passed to \code{print.birp}.}
}
\value{
Invisibly returns the input \code{object}, called for side effects.
}
\description{
Provides a printed summary of model estimates for a \code{birp} object.
}
\examples{
data <- simulate_birp()
est <- birp(data)
summary(est)

}
\seealso{
\code{\link{birp}}
}
