% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp_data.R
\name{summary.birp_data}
\alias{summary.birp_data}
\title{This function summarizes a birp_data object}
\usage{
\method{summary}{birp_data}(object, ...)
}
\arguments{
\item{object}{The birp_data object to be printed.}

\item{...}{Other parameters passed to function}
}
\value{
No return value, called for side effects
}
\description{
This function summarizes a birp_data object
}
\examples{
data <- simulate_birp()
summary(data)
}
