% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{birp_from_command_line}
\alias{birp_from_command_line}
\title{Create a birp Object from Command-Line Output Files}
\usage{
birp_from_command_line(path)
}
\arguments{
\item{path}{Character string specifying the directory path containing all birp output files.}
}
\value{
An object of type \code{birp} containing MCMC results and model estimates read from files.
}
\description{
This function creates a birp object by reading the output files generated by the command-line version of the birp tool.
}
\examples{
est <- birp_from_command_line(file.path(system.file("extdata", package = "birp")))
}
