% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{plot_mcmc}
\alias{plot_mcmc}
\title{Plotting the MCMC chains}
\usage{
plot_mcmc(x, col = c("black", "blue"))
}
\arguments{
\item{x}{A birp object}

\item{col}{Color(s) used in the plot}
}
\value{
No return value, called for side effects
}
\description{
Plotting the MCMC chains
}
\examples{
data <- simulate_birp()
est <- birp(data)
plot_mcmc(est)
}
\seealso{
\code{\link{birp}}
}
