% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp.R
\name{summary.birp}
\alias{summary.birp}
\title{Summarizing a birp object}
\usage{
\method{summary}{birp}(object, ...)
}
\arguments{
\item{object}{A birp object.}

\item{...}{Additional parameters passed to summary functions.}
}
\value{
No return value, called for side effects.
}
\description{
Summarizing a birp object
}
\examples{
data <- simulate_birp()
est <- birp(data)
summary(est)

}
\seealso{
\code{\link{birp}}
}
