% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birp_data.R
\name{summary.birp_data}
\alias{summary.birp_data}
\title{Summarize a birp_data object}
\usage{
\method{summary}{birp_data}(object, ...)
}
\arguments{
\item{object}{A \code{birp_data} object to be summarized.}

\item{...}{Additional arguments.}
}
\value{
No return value; this function is called for its side effects (printing to console).
}
\description{
Provides a printed summary of the contents of a \code{birp_data} object.
}
\examples{
data <- simulate_birp()
summary(data)
}
